Move PaymentOptionDisplayData class into CheckoutController.Session.
